refactor(front): extract report CSV and DOCX builders
